requirements.txt
# 依赖列表
requests==2.31.0
pytest==8.0.0
pytest-xdist==3.5.0
allure-pytest==2.13.2
pyyaml==6.0.1
faker==22.0.0
# 查看版本：pip show pymysql
pymysql==1.1.2